TAIS_EXAMEN/backend/gestion_productos/requirements.txt
setuptools
boto3
django-storages
zappa